beam

+++ mnmt/decoder/beam_decoder.py
@@ -41,7 +41,7 @@ class BeamDecoder(BasicDecoder):
 
         # Initialize the scores; for the first step,
         # ignore the inflated copies to avoid duplicate entries in the top k
-        sequence_scores = torch.tensor((batch_size * self.beam_size, 1)).to(self.device)
+        sequence_scores = torch.tensor((batch_size * self.beam_size, 1), dtype=torch.float64).to(self.device)
         sequence_scores.fill_(-float('Inf'))
         sequence_scores.index_fill_(0, torch.tensor([i * self.beam_size for i in range(0, batch_size)]), 0.0)
